refactor: type과 함수 로직을 분리하여 main.ts에서 조합하여 사용
