+++ index.ts
@@ -1,5 +1,1 @@
-export {
-  getTimePassed,
-  getTextFromTimePassed,
-  getDiff,
-} from './src/renderTime';
+export { getTimePassed, getTextFromTimePassed, getDiff } from './src/main';
