+++ src/renderTime.type.ts
@@ -1,0 +1,9 @@
+export type Time = {
+  id: string;
+  time: number;
+};
+
+export type RenderText = {
+  dateName: 'minute' | 'day' | 'hour' | 'month' | 'year';
+  timePassed: number;
+};
